+++ src/script/descriptor.h
@@ -7,6 +7,7 @@
 
 #include <script/script.h>
 #include <script/sign.h>
+#include <script/signingprovider.h>
 
 #include <vector>
