+++ src/wallet/ismine.cpp
@@ -8,6 +8,7 @@
 #include <key.h>
 #include <script/script.h>
 #include <script/sign.h>
+#include <script/signingprovider.h>
 #include <wallet/wallet.h>
 
 typedef std::vector<unsigned char> valtype;
